requirements.txt
alembic==1.4.0
aniso8601==8.0.0
astroid==2.3.3
attrs==20.2.0
base58==2.0.1
bcrypt==3.1.7
bitarray==1.2.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cytoolz==0.11.0
device-detector==0.10
dnspython==1.16.0
ecdsa==0.16.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
eventlet==0.25.1
expiringdict==1.2.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-HTTPAuth==3.3.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
greenlet==0.4.15
hexbytes==0.2.1
idna==2.9
imutils==0.5.3
ipfshttpclient==0.7.0a1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lru-dict==1.1.6
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
mccabe==0.6.1
monotonic==1.5
multiaddr==0.0.9
netaddr==0.8.0
numpy==1.18.1
parsimonious==0.8.1
passlib==1.7.2
Pillow==7.0.0
protobuf==3.13.0
pycparser==2.19
pycryptodome==3.9.9
PyJWT==1.7.1
pylint==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
PyMySQL==0.10.1
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.11.13
requests==2.23.0
rlp==2.0.0
rusty-rlp==0.1.15
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
toolz==0.11.1
twilio==6.48.0
typing==3.7.4.1
urllib3==1.25.8
varint==1.0.2
web3==5.13.0
websockets==8.1
Werkzeug==0.16.1
wrapt==1.11.2
